LayoutTests/platform/mac/accessibility/test-convenience-methods-expected.txt

This tests that test convenience methods behave correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS isFocusedConvenience is isFocusedConvenience
PASS isFocusedConvenience is isFocusedAttribute
PASS successfullyParsed is true

TEST COMPLETE
